requirements/dev.requirements.txt
alembic
asyncpg
fastapi
pydantic-settings
redis
sqlalchemy
uvicorn[standard]
celery
flower
rabbitmq
openpyxl
google-api-python-client
gspread-asyncio
httpx
